Move DataUpdateCoordinator to coordinator module in tesla_wall_connector
