+++ homeassistant/components/tesla_wall_connector/__init__.py
@@ -2,90 +2,38 @@
 
 from __future__ import annotations
 
-from dataclasses import dataclass
-from datetime import timedelta
-import logging
-
-from tesla_wall_connector import WallConnector
-from tesla_wall_connector.exceptions import (
-    WallConnectorConnectionError,
-    WallConnectorConnectionTimeoutError,
-    WallConnectorError,
-)
+from tesla_wall_connector.exceptions import WallConnectorError
 
 from homeassistant.config_entries import ConfigEntry
-from homeassistant.const import CONF_HOST, CONF_SCAN_INTERVAL, Platform
+from homeassistant.const import CONF_HOST, Platform
 from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import ConfigEntryNotReady
-from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator, UpdateFailed
 
-from .const import (
-    DEFAULT_SCAN_INTERVAL,
-    DOMAIN,
-    WALLCONNECTOR_DATA_LIFETIME,
-    WALLCONNECTOR_DATA_VITALS,
-)
+from .const import DOMAIN
+from .coordinator import WallConnectorCoordinator, WallConnectorData, get_poll_interval
 
 PLATFORMS: list[Platform] = [Platform.BINARY_SENSOR, Platform.SENSOR]
-
-_LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     """Set up Tesla Wall Connector from a config entry."""
     hass.data.setdefault(DOMAIN, {})
-    hostname = entry.data[CONF_HOST]
 
-    wall_connector = WallConnector(host=hostname, session=async_get_clientsession(hass))
+    coordinator = WallConnectorCoordinator(hass, entry)
 
     try:
-        version_data = await wall_connector.async_get_version()
+        version_data = await coordinator.wall_connector.async_get_version()
     except WallConnectorError as ex:
         raise ConfigEntryNotReady from ex
-
-    async def async_update_data():
-        """Fetch new data from the Wall Connector."""
-        try:
-            vitals = await wall_connector.async_get_vitals()
-            lifetime = await wall_connector.async_get_lifetime()
-        except WallConnectorConnectionTimeoutError as ex:
-            raise UpdateFailed(
-                f"Could not fetch data from Tesla WallConnector at {hostname}: Timeout"
-            ) from ex
-        except WallConnectorConnectionError as ex:
-            raise UpdateFailed(
-                f"Could not fetch data from Tesla WallConnector at {hostname}: Cannot"
-                " connect"
-            ) from ex
-        except WallConnectorError as ex:
-            raise UpdateFailed(
-                f"Could not fetch data from Tesla WallConnector at {hostname}: {ex}"
-            ) from ex
-
-        return {
-            WALLCONNECTOR_DATA_VITALS: vitals,
-            WALLCONNECTOR_DATA_LIFETIME: lifetime,
-        }
-
-    coordinator: DataUpdateCoordinator = DataUpdateCoordinator(
-        hass,
-        _LOGGER,
-        config_entry=entry,
-        name="tesla-wallconnector",
-        update_interval=get_poll_interval(entry),
-        update_method=async_update_data,
-    )
 
     await coordinator.async_config_entry_first_refresh()
 
     hass.data[DOMAIN][entry.entry_id] = WallConnectorData(
-        wall_connector_client=wall_connector,
-        hostname=hostname,
+        coordinator=coordinator,
+        hostname=entry.data[CONF_HOST],
         part_number=version_data.part_number,
         firmware_version=version_data.firmware_version,
         serial_number=version_data.serial_number,
-        update_coordinator=coordinator,
     )
 
     await hass.config_entries.async_forward_entry_setups(entry, PLATFORMS)
@@ -95,17 +43,10 @@ async def async_setup_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
     return True
 
 
-def get_poll_interval(entry: ConfigEntry) -> timedelta:
-    """Get the poll interval from config."""
-    return timedelta(
-        seconds=entry.options.get(CONF_SCAN_INTERVAL, DEFAULT_SCAN_INTERVAL)
-    )
-
-
 async def update_listener(hass: HomeAssistant, entry: ConfigEntry) -> None:
     """Handle options update."""
     wall_connector_data: WallConnectorData = hass.data[DOMAIN][entry.entry_id]
-    wall_connector_data.update_coordinator.update_interval = get_poll_interval(entry)
+    wall_connector_data.coordinator.update_interval = get_poll_interval(entry)
 
 
 async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
@@ -114,15 +55,3 @@ async def async_unload_entry(hass: HomeAssistant, entry: ConfigEntry) -> bool:
         hass.data[DOMAIN].pop(entry.entry_id)
 
     return unload_ok
-
-
-@dataclass
-class WallConnectorData:
-    """Data for the Tesla Wall Connector integration."""
-
-    wall_connector_client: WallConnector
-    update_coordinator: DataUpdateCoordinator
-    hostname: str
-    part_number: str
-    firmware_version: str
-    serial_number: str
