+++ homeassistant/components/tesla_wall_connector/binary_sensor.py
@@ -13,8 +13,8 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
 
-from . import WallConnectorData
 from .const import DOMAIN, WALLCONNECTOR_DATA_VITALS
+from .coordinator import WallConnectorData
 from .entity import WallConnectorEntity, WallConnectorLambdaValueGetterMixin
 
 _LOGGER = logging.getLogger(__name__)
